+++ sktime/forecasting/prophet.py
@@ -8,6 +8,7 @@ __all__ = ["Prophet"]
 
 from sktime.forecasting.base._adapters import _ProphetAdapter
 from sktime.utils.check_imports import _check_soft_dependencies
+from sktime.forecasting.base._base import DEFAULT_ALPHA
 
 
 _check_soft_dependencies("fbprophet")
@@ -84,13 +85,58 @@ class Prophet(_ProphetAdapter):
 
     """
 
-    def __init__(self, **kwargs):
-        super(Prophet, self).__init__(**kwargs)
+    def __init__(
+        self,
+        # Args due to wrapping
+        freq=None,
+        add_seasonality=None,
+        add_country_holidays=None,
+        # Args of fbprophet
+        growth="linear",
+        changepoints=None,
+        n_changepoints=25,
+        changepoint_range=0.8,
+        yearly_seasonality="auto",
+        weekly_seasonality="auto",
+        daily_seasonality="auto",
+        holidays=None,
+        seasonality_mode="additive",
+        seasonality_prior_scale=10.0,
+        holidays_prior_scale=10.0,
+        changepoint_prior_scale=0.05,
+        mcmc_samples=0,
+        interval_width=1 - DEFAULT_ALPHA,
+        uncertainty_samples=1000,
+        stan_backend=None,
+        **kwargs
+    ):
+        self.freq = freq
+        self.add_seasonality = add_seasonality
+        self.add_country_holidays = add_country_holidays
+
+        self.growth = growth
+        self.changepoints = changepoints
+        self.n_changepoints = n_changepoints
+        self.changepoint_range = changepoint_range
+        self.yearly_seasonality = yearly_seasonality
+        self.weekly_seasonality = weekly_seasonality
+        self.daily_seasonality = daily_seasonality
+        self.holidays = holidays
+        self.seasonality_mode = seasonality_mode
+        self.seasonality_prior_scale = float(seasonality_prior_scale)
+        self.changepoint_prior_scale = float(changepoint_prior_scale)
+        self.holidays_prior_scale = float(holidays_prior_scale)
+        self.mcmc_samples = mcmc_samples
+        self.interval_width = interval_width
+        self.uncertainty_samples = uncertainty_samples
+        self.stan_backend = stan_backend
 
         # import inside method to avoid hard dependency
         from fbprophet.forecaster import Prophet as _Prophet
 
         self._ModelClass = _Prophet
+
+        super(Prophet, self).__init__()
 
     def _instantiate_model(self):
         self._forecaster = self._ModelClass(
